requirements.txt
streamlit
pandas
numpy
opencv-python-headless
pytesseract
Pillow
XlsxWriter
keras
tensorflow
gdown
python-dotenv
matplotlib
torchvision
torch